+++ site/src/components/Dashboard/DashboardLayout.tsx
@@ -9,7 +9,7 @@ import { FC, Suspense } from "react"
 import { Outlet } from "react-router-dom"
 import { dashboardContentBottomPadding } from "theme/constants"
 import { updateCheckMachine } from "xServices/updateCheck/updateCheckXService"
-import { Navbar } from "../Navbar/Navbar"
+import { Navbar } from "./Navbar/Navbar"
 import Snackbar from "@mui/material/Snackbar"
 import Link from "@mui/material/Link"
 import Box, { BoxProps } from "@mui/material/Box"
