+++ site/src/components/Dashboard/Navbar/NavbarView.tsx
@@ -8,9 +8,9 @@ import { CoderIcon } from "components/Icons/CoderIcon"
 import { FC, useRef, useState } from "react"
 import { NavLink, useLocation, useNavigate } from "react-router-dom"
 import { colors } from "theme/colors"
-import * as TypesGen from "../../api/typesGenerated"
-import { navHeight } from "../../theme/constants"
-import { combineClasses } from "../../utils/combineClasses"
+import * as TypesGen from "../../../api/typesGenerated"
+import { navHeight } from "../../../theme/constants"
+import { combineClasses } from "../../../utils/combineClasses"
 import { UserDropdown } from "./UserDropdown/UserDropdown"
 import Box from "@mui/material/Box"
 import Menu from "@mui/material/Menu"
