+++ site/src/components/Dashboard/Navbar/UserDropdown/UserDropdown.stories.tsx
@@ -1,6 +1,6 @@
 import Box from "@mui/material/Box"
 import { Story } from "@storybook/react"
-import { MockUser } from "../../../testHelpers/entities"
+import { MockUser } from "../../../../testHelpers/entities"
 import { UserDropdown, UserDropdownProps } from "./UserDropdown"
 
 export default {
