+++ site/src/components/Dashboard/Navbar/UserDropdown/UserDropdown.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, screen } from "@testing-library/react"
-import { MockSupportLinks, MockUser } from "../../../testHelpers/entities"
-import { render } from "../../../testHelpers/renderHelpers"
+import { MockSupportLinks, MockUser } from "../../../../testHelpers/entities"
+import { render } from "../../../../testHelpers/renderHelpers"
 import { Language } from "./UserDropdownContent/UserDropdownContent"
 import { UserDropdown, UserDropdownProps } from "./UserDropdown"
