+++ site/src/components/UsersLayout/UsersLayout.tsx
@@ -3,7 +3,7 @@ import Link from "@mui/material/Link"
 import { makeStyles } from "@mui/styles"
 import GroupAdd from "@mui/icons-material/GroupAddOutlined"
 import PersonAdd from "@mui/icons-material/PersonAddOutlined"
-import { USERS_LINK } from "components/Navbar/NavbarView"
+import { USERS_LINK } from "components/Dashboard/Navbar/NavbarView"
 import { PageHeader, PageHeaderTitle } from "components/PageHeader/PageHeader"
 import { useFeatureVisibility } from "hooks/useFeatureVisibility"
 import { usePermissions } from "hooks/usePermissions"
